Moved the management pages into their own folder in refactor.
