+++ frontend/src/refactor/management_pages/ManagementPageWrappers.jsx
@@ -1,5 +1,5 @@
 import {useContext} from "react";
-import {UserContext} from "../context/userContext.jsx";
+import {UserContext} from "../../context/userContext.jsx";
 import ManagePeriod from "./ManagePeriod.jsx";
 
 export function ManagePeriodWrapper() {
